+++ Sources/Sentry/SentryHttpTransport.m
@@ -445,8 +445,11 @@
 - (void)recordLostSpans:(SentryEnvelopeItem *)envelopeItem reason:(SentryDiscardReason)reason
 {
     if ([SentryEnvelopeItemTypes.transaction isEqualToString:envelopeItem.header.type]) {
-        NSDictionary *_Nullable transactionJson =
-            [SentrySerialization deserializeDictionaryFromJsonData:envelopeItem.data];
+        if (envelopeItem.data == nil) {
+            return;
+        }
+        NSDictionary *_Nullable transactionJson = [SentrySerialization
+            deserializeDictionaryFromJsonData:SENTRY_UNWRAP_NULLABLE(NSData, envelopeItem.data)];
         if (transactionJson == nil) {
             return;
         }